Move conftest and fix import
