+++ tests/filesystem/test_filesystem.py
@@ -3,9 +3,9 @@ import os.path
 
 import pytest
 import trio.testing
-from datastore.core.test.conftest import DatastoreTests
 
 from datastore.filesystem import FileSystemDatastore
+from tests.conftest import DatastoreTests
 
 
 @pytest.fixture
